apply new "`categoryId`  workflow" in creating random transactions * fix adding refId to random transactions * update endpoint to delete all transactions by adding authentication and deleting only transactions of authenticated user * add endpoint to create random transactions when authenticated user doesn't have any transactions

+++ src/routes/transaction-routes/handlers/delete-all-transactions-handler.ts
@@ -1,8 +1,18 @@
+import { AppError } from "@utils/errors";
+import { UserModel } from "@models/user-model";
 import { FastifyReply, FastifyRequest } from "fastify";
+import { AuthenticatedRequest } from "@routes/routes-types";
 import { TransactionModel } from "@models/transaction-model";
 
 
 export const deleteAllTransactionsHandler = async (
-  _req: FastifyRequest,
+  req: FastifyRequest,
   res: FastifyReply,
-) => res.code(200).send(await TransactionModel.deleteMany());
+) => {
+  const ownerId = (req as AuthenticatedRequest).userId;
+  const user = await UserModel.findById(ownerId);
+  if (user?.email !== "test1@test.com")
+    throw new AppError(403, "Only one particular test user can delete its transactions");
+
+  return res.code(200).send(await TransactionModel.deleteMany({ ownerId }));
+}
